Create factory for dependent components
