+++ Assets/Scripts/Infrastructure/Factories/Component/DependentComponentFactory.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 95f00f70238242ecbfbc2d37c82aef40
+timeCreated: 1673373965
